+++ js/data.mjs
@@ -1,3 +1,4 @@
+import {getRandomArrayElement,getRandomInteger} from './util.mjs';
 const NAMES = [
   'Дмитрий',
   'Антон',
@@ -40,15 +41,6 @@ const avatarId = {
 };
 const POST_QUANTITY = 25;
 
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a,b));
-  const upper = Math.floor(Math.max(a,b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
 const createCommentId = () => {
   let currentId = 0;
   return () => ++currentId;
@@ -82,4 +74,4 @@ function generateComments(){
   };
 }
 
-generatePosts();
+export {generatePosts};
