+++ js/main.mjs
@@ -1,0 +1,5 @@
+import {generatePosts} from './data.mjs';
+
+console.log(
+  generatePosts()
+);
